Convert VolumeStream to use Vector3d's
There's effectively a problem with positions of blocks, they're
technically considered the corners of a dotted world, and since
transformations are performed on the double accuracy positioning, we
want to be able to retain those positions as much as possible throughout
however many transformations on a volume are made.

This is mostly to support the ability to rotate and mirror volumes in
their entirety. The implementations are already semi-functional, pending
 some off by one errors to be found and fixed.

+++ src/main/java/org/spongepowered/common/world/volume/buffer/archetype/blockentity/MutableMapBlockEntityArchetypeBuffer.java
@@ -110,7 +110,7 @@ public class MutableMapBlockEntityArchetypeBuffer extends AbstractMutableBlockEn
         }
         final Stream<VolumeElement<BlockEntityArchetypeVolume.Mutable, BlockEntityArchetype>> volumeElementStream = entryStream
             .filter(VolumeStreamUtils.blockEntityArchetypePositionFilter(min, max))
-            .map(entry -> VolumeElement.of(this, entry.getValue(), entry.getKey()));
+            .map(entry -> VolumeElement.of(this, entry.getValue(), entry.getKey().toDouble()));
 
         return new SpongeVolumeStream<>(volumeElementStream, () -> this);
     }
